feat(uv-tool): feat uv tool install by pyproject.toml
