+++ src/video_study_notes/__init__.py
@@ -1,0 +1,5 @@
+"""video-study-notes CLI package."""
+
+__all__ = ["__version__"]
+
+__version__ = "0.1.0"
